api/requirements.txt
djangorestframework
tensorflow
pillow
numpy
django